Refactor: Move Postgres container utilities to a centralized `postgres-test-utils` module to eliminate duplication across projects.
