+++ client-java/pom.xml
@@ -25,6 +25,7 @@
         <module>distance-heuristics</module>
         <module>test-utils-java</module>
         <module>test-old-libraries</module>
+        <module>postgres-test-utils</module>
     </modules>
 
     <licenses>
