+++ core-tests/e2e-tests/spring/spring-rest-postgres/pom.xml
@@ -116,6 +116,10 @@
             <artifactId>testcontainers</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.evomaster</groupId>
+            <artifactId>evomaster-client-java-postgres-test-utils</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
         </dependency>
