+++ core-tests/e2e-tests/spring/spring-rest-postgres/src/test/kotlin/com/foo/spring/rest/postgres/SpringRestPostgresController.kt
@@ -1,6 +1,7 @@
 package com.foo.spring.rest.postgres
 
 import org.evomaster.client.java.controller.EmbeddedSutController
+import org.evomaster.client.java.postgres.test.utils.PostgresContainerUtils
 import org.evomaster.client.java.controller.api.dto.auth.AuthenticationDto
 import org.evomaster.client.java.controller.api.dto.SutInfoDto
 import org.evomaster.client.java.controller.api.dto.database.schema.DatabaseType
