+++ core/pom.xml
@@ -23,6 +23,10 @@
         </dependency>
         <dependency>
             <groupId>org.evomaster</groupId>
+            <artifactId>evomaster-client-java-postgres-test-utils</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.evomaster</groupId>
             <artifactId>evomaster-test-utils-js</artifactId>
         </dependency>
         <dependency>
