+++ pom.xml
@@ -366,6 +366,11 @@
             </dependency>
             <dependency>
                 <groupId>org.evomaster</groupId>
+                <artifactId>evomaster-client-java-postgres-test-utils</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.evomaster</groupId>
                 <artifactId>evomaster-test-utils-js</artifactId>
                 <version>${project.version}</version>
             </dependency>
